pyproject.toml
[tool.poetry]
name = "fastapi-realworld-example-app"
version = "0.0.0"
description = "Backend logic implementation for https://github.com/gothinkster/realworld with awesome FastAPI"
authors = ["Nik Sidnev <sidnev.nick@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
uvicorn = "^0.18.2"
fastapi = "^0.79.1"
pydantic = { version = "^1.9", extras = ["email", "dotenv"] }
passlib = { version = "^1.7", extras = ["bcrypt"] }
pyjwt = "^2.4"
databases = "^0.6.1"
asyncpg = "^0.26.0"
psycopg2-binary = "^2.9.3"
aiosql = "^6.2"
pypika = "^0.48.9"
alembic = "^1.8"
python-slugify = "^6.1"
Unidecode = "^1.3"
loguru = "^0.6.0"
langchain-openai = "^0.3.15"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
isort = "^5.10"
autoflake = "^1.4"
wemake-python-styleguide = "^0.16.1"
mypy = "^0.971"
flake8-fixme = "^1.1"
pytest = "^7.1"
pytest-cov = "^3.0"
pytest-asyncio = "^0.19.0"
pytest-env = "^0.6.2"
pytest-xdist = "^2.4.0"
httpx = "^0.23.0"
asgi-lifespan = "^1.0.1"

[tool.isort]
profile = "black"
src_paths = ["app", "tests"]
combine_as_imports = true

[tool.pytest.ini_options]
testpaths = "tests"
filterwarnings = "error"
addopts = '''
  --strict-markers
  --tb=short
  --cov=app
  --cov=tests
  --cov-branch
  --cov-report=term-missing
  --cov-report=html
  --cov-report=xml
  --no-cov-on-fail
  --cov-fail-under=100
  --numprocesses=auto
  --asyncio-mode=auto
'''
env = [
  "SECRET_KEY=secret",
  "MAX_CONNECTIONS_COUNT=1",
  "MIN_CONNECTIONS_COUNT=1"
]

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"
